Move security_hal.h to external/include/security/
- Included in multiple applications
