+++ external/include/security/hal/security_hal.h
