+++ os/se/sss/security_hal_wrapper.c
@@ -20,7 +20,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "security_hal.h"
+#include <security/hal/security_hal.h>
 #include "isp_custom.h"
 
 #define ISP_CHECKBUSY() while (isp_get_status()) {}
